pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embodiedai-core"
version = "0.1.0"
description = "具身智能核心框架 | Embodied AI Core Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "EmbodiedAI-Core Contributors", email = "embodiedai@example.com"}
]
keywords = [
    "embodied-ai",
    "robotics",
    "vision-language-action",
    "vla",
    "robot-control",
    "simulation",
    "pytorch",
    "computer-vision"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pybullet>=3.2.0",
    "mujoco>=2.3.0",
    "ompl>=1.6.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]
simulation = [
    "isaacgym",
    "gymnasium>=0.29.0",
]

[project.urls]
Homepage = "https://github.com/ENDcodeworld/EmbodiedAI-Core"
Documentation = "https://github.com/ENDcodeworld/EmbodiedAI-Core/tree/main/docs"
Repository = "https://github.com/ENDcodeworld/EmbodiedAI-Core"
"Bug Tracker" = "https://github.com/ENDcodeworld/EmbodiedAI-Core/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["embodiedai*"]

[tool.setuptools.package-data]
embodiedai = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=embodiedai --cov-report=term-missing"
